moving RandomAccess/Indexable class&instances
