+++ basement/Basement/BoxedArray.hs
@@ -82,7 +82,8 @@ import           Basement.Numerical.Additive
 import           Basement.Numerical.Subtractive
 import           Basement.NonEmpty
 import           Basement.Compat.Base
-import qualified Basement.Alg.Mutable as MutAlg
+import qualified Basement.Alg.Class as Alg
+import qualified Basement.Alg.Mutable as Alg
 import           Basement.Compat.MonadTrans
 import           Basement.Compat.Semigroup
 import           Basement.Types.OffsetSize
@@ -631,7 +632,7 @@ find predicate vec = loop 0
              in if predicate e then Just e else loop (i+1)
 
 instance (PrimMonad prim, st ~ PrimState prim) 
-         => MutAlg.RandomAccess (MArray ty st) prim ty where
+         => Alg.RandomAccess (MArray ty st) prim ty where
     read (MArray _ _ mba) = primMutableArrayRead mba
     write (MArray _ _ mba) = primMutableArrayWrite mba
 
@@ -642,7 +643,7 @@ sortBy xford vec
   where
     len = length vec
     doSort :: PrimMonad prim => (ty -> ty -> Ordering) -> MArray ty (PrimState prim) -> prim (Array ty)
-    doSort ford ma = MutAlg.inplaceSortBy ford 0 len ma >> unsafeFreeze ma
+    doSort ford ma = Alg.inplaceSortBy ford 0 len ma >> unsafeFreeze ma
 
 filter :: forall ty . (ty -> Bool) -> Array ty -> Array ty
 filter predicate vec = runST (new len >>= copyFilterFreeze predicate (unsafeIndex vec))
